requirements.txt
flask
pandas
numpy
scikit-learn
python-dotenv
groq
